requirements.txt
pickledb
discord.py